requirements.txt
#Basic
pandas
numpy

#Model
transformers
torch

#Gcloud
google-cloud
google-cloud-bigquery
db-dtypes

#API
fastapi
uvicorn
python-multipart
pillow

#streamlit
streamlit